Extract useCurrentPath for stable popstate coverage
Move browser history listening into a dedicated hook with a focused
unit test so ReactApp coverage stays at 100% without flaky title checks.
