+++ modules/core/client/react-app/useCurrentPath.js
@@ -1,0 +1,15 @@
+import { useEffect, useState } from 'react';
+
+export function useCurrentPath() {
+  const [currentPath, setCurrentPath] = useState(window.location.pathname);
+
+  useEffect(() => {
+    const onPopState = () => setCurrentPath(window.location.pathname);
+
+    window.addEventListener('popstate', onPopState);
+
+    return () => window.removeEventListener('popstate', onPopState);
+  }, []);
+
+  return currentPath;
+}
